Address plugin refactor review comments
