+++ codex-rs/config/src/lib.rs
@@ -14,6 +14,7 @@ mod mcp_types;
 mod merge;
 mod overrides;
 pub mod permissions_toml;
+mod plugin_edit;
 pub mod profile_toml;
 mod project_root_markers;
 mod requirements_exec_policy;
