+++ codex-rs/core-api/Cargo.toml
@@ -17,6 +17,7 @@ codex-arg0 = { workspace = true }
 codex-analytics = { workspace = true }
 codex-config = { workspace = true }
 codex-core = { workspace = true }
+codex-core-plugins = { workspace = true }
 codex-exec-server = { workspace = true }
 codex-features = { workspace = true }
 codex-login = { workspace = true }
