+++ codex-rs/core/src/agent/role_tests.rs
@@ -2,10 +2,9 @@ use super::*;
 use crate::SkillsManager;
 use crate::config::CONFIG_TOML_FILE;
 use crate::config::ConfigBuilder;
-use crate::plugins::PluginsManager;
 use crate::skills_load_input_from_config;
 use codex_config::ConfigLayerStackOrdering;
-use codex_features::Feature;
+use codex_core_plugins::PluginsManager;
 use codex_protocol::config_types::ReasoningSummary;
 use codex_protocol::config_types::Verbosity;
 use codex_protocol::openai_models::ReasoningEffort;
@@ -657,12 +656,7 @@ enabled = false
     let skills_manager =
         SkillsManager::new(home.path().abs(), /*bundled_skills_enabled*/ true);
     let plugin_outcome = plugins_manager
-        .plugins_for_config(
-            &config.config_layer_stack,
-            config.features.enabled(Feature::Plugins),
-            config.features.enabled(Feature::RemotePlugin),
-            config.features.enabled(Feature::PluginHooks),
-        )
+        .plugins_for_config(&config.config_layer_stack, config.plugin_feature_flags())
         .await;
     let effective_skill_roots = plugin_outcome.effective_skill_roots();
     let skills_input = skills_load_input_from_config(&config, effective_skill_roots);
