+++ codex-rs/core/src/mcp.rs
@@ -2,8 +2,8 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use crate::config::Config;
-use crate::plugins::PluginsManager;
 use codex_config::McpServerConfig;
+use codex_core_plugins::PluginsManager;
 use codex_login::CodexAuth;
 use codex_mcp::ToolPluginProvenance;
 use codex_mcp::configured_mcp_servers;
