+++ codex-rs/core/src/mcp_tool_call.rs
@@ -758,9 +758,7 @@ async fn custom_mcp_tool_approval_mode(
         .plugins_manager
         .plugins_for_config(
             &turn_context.config.config_layer_stack,
-            turn_context.config.features.enabled(Feature::Plugins),
-            turn_context.config.features.enabled(Feature::RemotePlugin),
-            turn_context.config.features.enabled(Feature::PluginHooks),
+            turn_context.config.plugin_feature_flags(),
         )
         .await
         .plugins()
@@ -1810,12 +1808,7 @@ async fn persist_non_app_mcp_tool_approval(
     let plugin_config_name = sess
         .services
         .plugins_manager
-        .plugins_for_config(
-            &config.config_layer_stack,
-            config.features.enabled(Feature::Plugins),
-            config.features.enabled(Feature::RemotePlugin),
-            config.features.enabled(Feature::PluginHooks),
-        )
+        .plugins_for_config(&config.config_layer_stack, config.plugin_feature_flags())
         .await
         .plugins()
         .iter()
