+++ codex-rs/core/src/mcp_tool_call_tests.rs
@@ -13,7 +13,6 @@ use codex_config::types::ApprovalsReviewer;
 use codex_config::types::AppsConfigToml;
 use codex_config::types::McpServerConfig;
 use codex_config::types::McpServerToolConfig;
-use codex_features::Feature;
 use codex_hooks::Hooks;
 use codex_hooks::HooksConfig;
 use codex_model_provider::create_model_provider;
@@ -1597,9 +1596,7 @@ enabled = true
         .plugins_manager
         .plugins_for_config(
             &initial_config.config_layer_stack,
-            initial_config.features.enabled(Feature::Plugins),
-            initial_config.features.enabled(Feature::RemotePlugin),
-            initial_config.features.enabled(Feature::PluginHooks),
+            initial_config.plugin_feature_flags(),
         )
         .await;
     std::fs::write(
