+++ codex-rs/core/src/session/handlers.rs
@@ -23,7 +23,6 @@ use codex_config::CloudRequirementsLoader;
 use codex_config::LoaderOverrides;
 use codex_config::loader::load_config_layers_state;
 use codex_exec_server::LOCAL_FS;
-use codex_features::Feature;
 use codex_utils_absolute_path::AbsolutePathBuf;
 
 use crate::review_prompts::resolve_review_request;
@@ -624,9 +623,7 @@ pub async fn list_skills(sess: &Session, sub_id: String, cwds: Vec<PathBuf>, for
         let effective_skill_roots = plugins_manager
             .effective_skill_roots_for_layer_stack(
                 &config_layer_stack,
-                config.features.enabled(Feature::Plugins),
-                config.features.enabled(Feature::RemotePlugin),
-                config.features.enabled(Feature::PluginHooks),
+                config.plugin_feature_flags(),
             )
             .await;
         let skills_input = crate::SkillsLoadInput::new(
