+++ codex-rs/core/src/session/tests.rs
@@ -3560,9 +3560,7 @@ pub(crate) async fn make_session_and_context() -> (Session, TurnContext) {
         .plugins_manager
         .plugins_for_config(
             &per_turn_config.config_layer_stack,
-            per_turn_config.features.enabled(Feature::Plugins),
-            per_turn_config.features.enabled(Feature::RemotePlugin),
-            per_turn_config.features.enabled(Feature::PluginHooks),
+            per_turn_config.plugin_feature_flags(),
         )
         .await;
     let effective_skill_roots = plugin_outcome.effective_skill_roots();
@@ -4994,9 +4992,7 @@ where
         .plugins_manager
         .plugins_for_config(
             &per_turn_config.config_layer_stack,
-            per_turn_config.features.enabled(Feature::Plugins),
-            per_turn_config.features.enabled(Feature::RemotePlugin),
-            per_turn_config.features.enabled(Feature::PluginHooks),
+            per_turn_config.plugin_feature_flags(),
         )
         .await;
     let effective_skill_roots = plugin_outcome.effective_skill_roots();
